pyproject.toml
[project]
name = "ai-trading-agent-python"
version = "0.1.0"
description = "AI Trading Agent using Lighter SDK and Claude"
requires-python = ">=3.10"
dependencies = [
    "anthropic>=0.39.0",
    "sqlalchemy[asyncio]>=2.0.0",
    "asyncpg>=0.29.0",
    "fastapi>=0.115.0",
    "uvicorn>=0.32.0",
    "httpx>=0.27.0",
    "pydantic>=2.0.0",
    "python-dotenv>=1.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.24.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"


